pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "strands-ai-functions"
dynamic = ["version"]
description = "AI Function decorator for enhancing functions with AI capabilities"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "AWS", email = "opensource@amazon.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "strands-agents>=1.16.0",
    "pydantic>=2.4.0,<3.0.0",
    "smolagents>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/strands-labs/ai-functions"
Documentation = "https://github.com/strands-labs/ai-functions#readme"
Repository = "https://github.com/strands-labs/ai-functions.git"
Issues = "https://github.com/strands-labs/ai-functions/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_functions"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "pytest --cov=ai_functions {args:tests}"
lint = "ruff check src tests"
format = "ruff format src tests"
typecheck = "mypy src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
include = ["examples/**/*.py", "src/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
select = ["B", "D", "E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"!src/**/*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
follow_untyped_imports = true
ignore_missing_imports = false

[tool.uv.workspace]
members = ["examples"]